--- pytest_helm_charts/giantswarm_app_platform/custom_resources.py
@@ -1,15 +1,0 @@
-from pykube.objects import APIObject, NamespacedAPIObject
-from deprecated import deprecated
-
-
-@deprecated(version="0.5.3", reason="Please use `CatalogCR` instead.")
-class AppCatalogCR(APIObject):
-    version = "application.giantswarm.io/v1alpha1"
-    endpoint = "appcatalogs"
-    kind = "AppCatalog"
-
-
-class CatalogCR(NamespacedAPIObject):
-    version = "application.giantswarm.io/v1alpha1"
-    endpoint = "catalogs"
-    kind = "Catalog"
